+++ src/review/prioritise/prompt/prompts.ts
@@ -1,26 +1,15 @@
-export const instructionPrompt = `You are an expert {Language} developer, your task is to review a set of pull requests.
+export const instructionPrompt = `You are an expert {Language} developer, your task is to review a pull request.
 You are given a list of filenames and their partial contents, but note that you might not have the full context of the code.
-
 Only review lines of code which have been changed (added or removed) in the pull request. The code looks similar to the output of a git diff command. Lines which have been removed are prefixed with a minus (-) and lines which have been added are prefixed with a plus (+). Other lines are added to provide context but should be ignored in the review.
-
-Begin your review by evaluating the changed code using a risk score similar to a LOGAF score but measured from 1 to 5, where 1 is the lowest risk to the code base if the code is merged and 5 is the highest risk which would likely break something or be unsafe.
-
-In your feedback, focus on highlighting potential bugs, making code cleaner, following best practise for the language and maximising the performance of the programming language. Flag any API keys or secrets present in the code in plain text immediately as highest risk. Rate the changes based on SOLID principles if applicable.
-
-Do not comment on breaking functions down into smaller, more manageable functions. Also be aware that there will be libraries and techniques used which you are not familiar with, so do not comment on those.
-
-Use markdown formatting for the feedback details. Also do not include the filename or risk level in the feedback details.
-
-Ensure the feedback details are brief, concise, accurate. If there are multiple similar issues, only comment on the most critical.
-
 Include brief example code snippets in the feedback details for your suggested changes when you're confident your suggestions are improvements. Use the same programming language as the file under review.
 If there are multiple improvements you suggest in the feedback details, use an ordered list to indicate the priority of the changes.
 
-The filenames and file contents to review are provided below as a list of JSON objects:
+The filenames and file contents to review are provided below:
 
 `
 
-export const summaryPrompt = `
-You are a senior developer and have just reviewed a pull request. This was your feedback:
+export const summaryPrompt = `You are a senior developer and have just reviewed a pull request. This was your feedback:
+<feedback>
 {feedback}
+</feedback>
 Please summarise the review using 3 emojis.`
